refactor(maz-ui): use new path to lazy img directive
